requirements.txt
disnake
mysql.connector
requests
flask